+++ gridstatus/nyiso.py
@@ -314,7 +314,11 @@ class NYISO(ISOBase):
 
         df = df.rename(columns=columns)
 
-        df["Energy"] = df["LMP"] - (df["Loss"] - df["Congestion"])
+        # In NYISO raw data, a negative congestion number means a higher LMP. We
+        # flip the sign to make it consistent with other ISOs where a negative
+        # congestion number means a lower LMP.
+        df["Congestion"] *= -1
+        df["Energy"] = df["LMP"] - df["Loss"] - df["Congestion"]
         df["Market"] = market.value
         df["Location Type"] = "Zone" if location_type == ZONE else "Generator"